all: Update directory structure of lexers and parsers.
To facilitate the comparison of hand-written and machine generated
lexers and parsers, this commit enforces a consistent directory
structure for such packages.

Hand-written packages will be contained within the "hand" directory,
while machine-generated packages will be contained within directories
of the given tool names.

The intended directory structure is along the lines of.

* gocc/lexer
* gocc/parser
* hand/lexer
* hand/parser
* hand/token
* lex/lexer
* peg/parser
* ragel/lexer
* yacc/parser