hand/lexer: Add long-char.c test case.
Not quite sure what the best approach is for error detection here.
Trying to be cleaver to potentially give better error messages, or be
simple but consistent and fail early for invalid tokens.

Approach 1)

If trying to be cleaver, where is the cut-off point? 1 byte look-ahead,
to locate 'cc' as invalid, 2 byte look-ahead, to also identify 'ccc',
etc.

Approach 2)

Another approach is to backtrack to the first possible starting position
of a valid token within the input byte stream. In the case of `'cc';`,
the lexer would emit an "unterminated character literal" error for the
first apostrophe and then backtrack to the position directly succeeding
the first apostrophe (i.e. `cc';`) to continue lexing. The lexer would
then emit the identifier `cc`, an error for the second unterminated
apostrophe and a semicolon token.

Approach 3) (current approach)

Lex as many bytes as would be valid for the current token being lexed.
In the case of `'cc'; // Not OK` being lexed, lex `'c` before emitting
an "unterminated character literal" error, as the first apostrophe
indicates that a character literal is to be lexed, the first `c`
indicates a valid character was located within the character literal,
and as the lexer tries to locate the terminating apostrophe, but fails
to do so, it emits an error, and continues lexing from the next byte in
the byte stream, i.e. from `c'; // Not OK`. Which would be lexed as the
identifier `c`, an "unterminated character literal" error for `';`, and
a comment for `// Not OK`

@sangisos any idea on which approach that is preferable? What are the
advantages and disadvantages with the three different approaches? Are
there any other approaches we may try?