gocc/grammar: Ignore comments. Remove special EOF rule for line comments.
Before the change, the following error occurs when parsing l03.c

    u@x1 ~/D/g/s/g/m/uc> uparse testdata/quiet/lexer/l03.c

    ...

    pos: 43
    typ: 33
    lit: // Was:   i = 1234567890;

    2016/04/04 13:20:46 main.parseFile (uparse.go:71): error: Error in S82: comment(33,// Was:   i = 1234567890;
    ), Pos(offset=43, line=6, column=2), expected one of: ; ident ( int_lit { } return while if - !

Ignoring comments but keeping the special EOF rule for line comments, results
in the following behaviour when lexing (and parsing) l03.c.

    u@x1 ~/D/g/s/g/m/uc> uparse testdata/quiet/lexer/l03.c

    pos: 46
    typ: 4
    lit: Was

    S82 ident(4,Was) reduce:25(Stmt : Expr ";"	<<  >>)
    S47 ident(4,Was) shift:42
    pos: 49
    typ: 0
    lit: :

    2016/04/04 13:22:54 main.parseFile (uparse.go:71): error: Error in S42: INVALID(0,:), Pos(offset=49, line=5, column=30), expected one of: ; ( [ - * / + < > <= >= == != && =

The rule

    _line_comment : '/' '/' { . } ; // line comment ending with EOF

is interpreted to lex the shortest prefix "// " before another valid lexeme,
which is not intended. The special EOF rule was always a hack, and a better
solution should be identified. Perhaps appending a newline to the file before
lexing the input?

In either case, removing the special EOF line comment rule, and ignoring
comments in the grammar solves two real issues discovered when parsing.